+++ inc/form_element_form.php
@@ -107,9 +107,16 @@ class form_element_form extends form_element {
     $table=$document->createElement("table");
     $table->setAttribute("id", $this->id);
     $table->setAttribute("class", "form");
+    $element_list = array();
 
     foreach($this->elements as $k=>$element) {
-      $table->appendChild($element->show($document));
+      $element_list[] = array($element->weight(), $element->show($document));
+    }
+
+    $element_list = weight_sort($element_list);
+
+    foreach($element_list as $element) {
+      $table->appendChild($element);
     }
 
     return $table;